Forced variable to be an integer
I added a snippet that forces j in space_loop to be an integer in order
to force the range to be between integers. Even through this range is
between integers, python still considers it not to be between two
integers. I plan to replace the for loop with a while loop.