Bugs Fixed, Code Runs
I am still not sure why the for loop in the function “space_loop” was
not counting up, but I replaced it with a while loop, and connected the
outer part of the upwind scheme in the function “time_loop” by setting
oldgrid = newgrid (ln35) in between time steps. The code appear to
generate a plausible answer, but I haven’t checked how accurate it is.
My next goal will be to make this code general for a number of initial
conditions.